Support DRBD path
A DRBD device can only be used by one process at the same time, so only one file descriptor can be opened.
Unfortunately xhad opens one descriptor per host on the heartbeat VDI. In the case of LINSTOR this mechanism
does not work with DRBD VDIs, the solution in this case is to open a descriptor only when a read or write is executed
(the descriptor is closed after the request). If another host tries to do the same thing, we put the daemon to sleep.

Signed-off-by: Ronan Abhamon <ronan.abhamon@vates.tech>